+++ src/main/java/org/apache/datasketches/kll/KllFloatsHelper.java
@@ -19,6 +19,8 @@
 
 package org.apache.datasketches.kll;
 
+import static java.lang.Math.max;
+import static java.lang.Math.min;
 import static org.apache.datasketches.Util.isEven;
 import static org.apache.datasketches.Util.isOdd;
 
@@ -32,7 +34,281 @@ import org.apache.datasketches.SketchesArgumentException;
  * @author Kevin Lang
  * @author Alexander Saydakov
  */
-class KllFloatsHelper {
+final class KllFloatsHelper {
+
+  static double getFloatRank(final KllSketch mine, final float value) {
+    if (mine.isEmpty()) { return Double.NaN; }
+    int level = 0;
+    int weight = 1;
+    long total = 0;
+    final float[] myFloatItemsArr = mine.getFloatItemsArray();
+    final int[] myLevelsArr = mine.getLevelsArray();
+    while (level < mine.getNumLevels()) {
+      final int fromIndex = myLevelsArr[level];
+      final int toIndex = myLevelsArr[level + 1]; // exclusive
+      for (int i = fromIndex; i < toIndex; i++) {
+        if (myFloatItemsArr[i] < value) {
+          total += weight;
+        } else if (level > 0 || mine.isLevelZeroSorted()) {
+          break; // levels above 0 are sorted, no point comparing further
+        }
+      }
+      level++;
+      weight *= 2;
+    }
+    return (double) total / mine.getN();
+  }
+
+  static double[] getFloatsPmfOrCdf(final KllSketch mine, final float[] splitPoints, final boolean isCdf) {
+    if (mine.isEmpty()) { return null; }
+    validateFloatValues(splitPoints);
+    final double[] buckets = new double[splitPoints.length + 1];
+    final int myNumLevels = mine.getNumLevels();
+    final int[] myLevelsArr = mine.getLevelsArray();
+    int level = 0;
+    int weight = 1;
+    while (level < myNumLevels) {
+      final int fromIndex = myLevelsArr[level];
+      final int toIndex = myLevelsArr[level + 1]; // exclusive
+      if (level == 0 && !mine.isLevelZeroSorted()) {
+        KllFloatsHelper.incrementFloatBucketsUnsortedLevel(mine, fromIndex, toIndex, weight, splitPoints, buckets);
+      } else {
+        KllFloatsHelper.incrementFloatBucketsSortedLevel(mine, fromIndex, toIndex, weight, splitPoints, buckets);
+      }
+      level++;
+      weight *= 2;
+    }
+    // normalize and, if CDF, convert to cumulative
+    if (isCdf) {
+      double subtotal = 0;
+      for (int i = 0; i < buckets.length; i++) {
+        subtotal += buckets[i];
+        buckets[i] = subtotal / mine.getN();
+      }
+    } else {
+      for (int i = 0; i < buckets.length; i++) {
+        buckets[i] /= mine.getN();
+      }
+    }
+    return buckets;
+  }
+
+  static float getFloatsQuantile(final KllSketch mine, final double fraction) {
+    if (mine.isEmpty()) { return Float.NaN; }
+    if (fraction < 0.0 || fraction > 1.0) {
+      throw new SketchesArgumentException("Fraction cannot be less than zero nor greater than 1.0");
+    }
+    //These two assumptions make KLL compatible with the previous classic Quantiles Sketch
+    if (fraction == 0.0) { return mine.getMinFloatValue(); }
+    if (fraction == 1.0) { return mine.getMaxFloatValue(); }
+    final KllFloatsQuantileCalculator quant = KllFloatsHelper.getFloatsQuantileCalculator(mine);
+    return quant.getQuantile(fraction);
+  }
+
+  static float[] getFloatsQuantiles(final KllSketch mine, final double[] fractions) {
+    if (mine.isEmpty()) { return null; }
+    KllFloatsQuantileCalculator quant = null;
+    final float[] quantiles = new float[fractions.length];
+    for (int i = 0; i < fractions.length; i++) {
+      final double fraction = fractions[i];
+      if (fraction < 0.0 || fraction > 1.0) {
+        throw new SketchesArgumentException("Fraction cannot be less than zero nor greater than 1.0");
+      }
+      if      (fraction == 0.0) { quantiles[i] = mine.getMinFloatValue(); }
+      else if (fraction == 1.0) { quantiles[i] = mine.getMaxFloatValue(); }
+      else {
+        if (quant == null) {
+          quant = KllFloatsHelper.getFloatsQuantileCalculator(mine);
+        }
+        quantiles[i] = quant.getQuantile(fraction);
+      }
+    }
+    return quantiles;
+  }
+
+  static void mergeFloatImpl(final KllSketch mine, final KllSketch other) {
+    if (other.isEmpty()) { return; }
+    final long finalN = mine.getN() + other.getN();
+    //update this sketch with level0 items from the other sketch
+    final float[] otherFloatItemsArr = other.getFloatItemsArray();
+    final int otherNumLevels = other.getNumLevels();
+    final int[] otherLevelsArr = other.getLevelsArray();
+    for (int i = otherLevelsArr[0]; i < otherLevelsArr[1]; i++) {
+      KllFloatsHelper.updateFloat(mine, otherFloatItemsArr[i]);
+    }
+    // after the level 0 update, we capture the key mutable variables
+    final float myMin = mine.getMinFloatValue();
+    final float myMax = mine.getMaxFloatValue();
+    final int myMinK = mine.getMinK();
+
+    final int myCurNumLevels = mine.getNumLevels();
+    final int[] myCurLevelsArr = mine.getLevelsArray();
+    final float[] myCurFloatItemsArr = mine.getFloatItemsArray();
+
+    final int myNewNumLevels;
+    final int[] myNewLevelsArr;
+    final float[] myNewFloatItemsArr;
+
+    if (otherNumLevels > 1) { //now merge higher levels if they exist
+      final int tmpSpaceNeeded = mine.getNumRetained()
+          + KllHelper.getNumRetainedAboveLevelZero(otherNumLevels, otherLevelsArr);
+      final float[] workbuf = new float[tmpSpaceNeeded];
+      final int ub = KllHelper.ubOnNumLevels(finalN);
+      final int[] worklevels = new int[ub + 2]; // ub+1 does not work
+      final int[] outlevels  = new int[ub + 2];
+
+      final int provisionalNumLevels = max(myCurNumLevels, otherNumLevels);
+
+      populateFloatWorkArrays(mine, other, workbuf, worklevels, provisionalNumLevels);
+
+      // notice that workbuf is being used as both the input and output
+      final int[] result = generalFloatsCompress(mine.getK(), mine.getM(), provisionalNumLevels,
+          workbuf, worklevels, workbuf, outlevels, mine.isLevelZeroSorted(), KllSketch.random);
+      final int targetItemCount = result[1]; //was finalCapacity. Max size given k, m, numLevels
+      final int curItemCount = result[2]; //was finalPop
+
+      // now we need to finalize the results for the "self" sketch
+
+      //THE NEW NUM LEVELS
+      myNewNumLevels = result[0]; //was finalNumLevels
+      assert myNewNumLevels <= ub; // ub may be much bigger
+
+      // THE NEW ITEMS ARRAY (was newbuf)
+      myNewFloatItemsArr = (targetItemCount == myCurFloatItemsArr.length)
+          ? myCurFloatItemsArr
+          : new float[targetItemCount];
+      final int freeSpaceAtBottom = targetItemCount - curItemCount;
+      //shift the new items array
+      System.arraycopy(workbuf, outlevels[0], myNewFloatItemsArr, freeSpaceAtBottom, curItemCount);
+      final int theShift = freeSpaceAtBottom - outlevels[0];
+
+      //calculate the new levels array length
+      final int finalLevelsArrLen;
+      if (myCurLevelsArr.length < myNewNumLevels + 1) { finalLevelsArrLen = myNewNumLevels + 1; }
+      else { finalLevelsArrLen = myCurLevelsArr.length; }
+
+      //THE NEW LEVELS ARRAY
+      myNewLevelsArr = new int[finalLevelsArrLen];
+      for (int lvl = 0; lvl < myNewNumLevels + 1; lvl++) { // includes the "extra" index
+        myNewLevelsArr[lvl] = outlevels[lvl] + theShift;
+      }
+
+      //MEMORY SPACE MANAGEMENT
+      if (mine.updatablMemory) {
+        mine.wmem = KllHelper.memorySpaceMgmt(mine, myNewLevelsArr.length, myNewFloatItemsArr.length);
+      }
+
+    } else {
+      myNewNumLevels = myCurNumLevels;
+      myNewLevelsArr = myCurLevelsArr;
+      myNewFloatItemsArr = myCurFloatItemsArr;
+    }
+
+    //Update Preamble:
+    mine.setN(finalN);
+    if (other.isEstimationMode()) { //otherwise the merge brings over exact items.
+      mine.setMinK(min(myMinK, other.getMinK()));
+    }
+
+    //Update min, max values
+    final float otherMin = other.getMinFloatValue();
+    final float otherMax = other.getMaxFloatValue();
+    mine.setMinFloatValue(resolveFloatMinValue(myMin, otherMin));
+    mine.setMaxFloatValue(resolveFloatMaxValue(myMax, otherMax));
+
+    //Update numLevels, levelsArray, items
+    mine.setNumLevels(myNewNumLevels);
+    mine.setLevelsArray(myNewLevelsArr);
+    mine.setFloatItemsArray(myNewFloatItemsArr);
+    assert KllHelper.sumTheSampleWeights(mine.getNumLevels(), mine.getLevelsArray()) == mine.getN();
+  }
+
+  static void mergeSortedFloatArrays(
+      final float[] bufA, final int startA, final int lenA,
+      final float[] bufB, final int startB, final int lenB,
+      final float[] bufC, final int startC) {
+    final int lenC = lenA + lenB;
+    final int limA = startA + lenA;
+    final int limB = startB + lenB;
+    final int limC = startC + lenC;
+
+    int a = startA;
+    int b = startB;
+
+    for (int c = startC; c < limC; c++) {
+      if (a == limA) {
+        bufC[c] = bufB[b];
+        b++;
+      } else if (b == limB) {
+        bufC[c] = bufA[a];
+        a++;
+      } else if (bufA[a] < bufB[b]) {
+        bufC[c] = bufA[a];
+        a++;
+      } else {
+        bufC[c] = bufB[b];
+        b++;
+      }
+    }
+    assert a == limA;
+    assert b == limB;
+  }
+
+  /**
+   * Validation Method. This must be modified to test validation
+   * @param buf the items array
+   * @param start data start
+   * @param length items length
+   * @param random instance of Random
+   */
+  static void randomlyHalveDownFloats(final float[] buf, final int start, final int length, final Random random) {
+    assert isEven(length);
+    final int half_length = length / 2;
+    final int offset = random.nextInt(2);       // disable for validation
+    //final int offset = deterministicOffset(); // enable for validation
+    int j = start + offset;
+    for (int i = start; i < (start + half_length); i++) {
+      buf[i] = buf[j];
+      j += 2;
+    }
+  }
+
+  /**
+   * Validation Method. This must be modified to test validation
+   * @param buf the items array
+   * @param start data start
+   * @param length items length
+   * @param random instance of Random
+   */
+  static void randomlyHalveUpFloats(final float[] buf, final int start, final int length, final Random random) {
+    assert isEven(length);
+    final int half_length = length / 2;
+    final int offset = random.nextInt(2);       // disable for validation
+    //final int offset = deterministicOffset(); // enable for validation
+    int j = (start + length) - 1 - offset;
+    for (int i = (start + length) - 1; i >= (start + half_length); i--) {
+      buf[i] = buf[j];
+      j -= 2;
+    }
+  }
+
+  static void updateFloat(final KllSketch mine, final float value) {
+    if (Float.isNaN(value)) { return; }
+    if (mine.isEmpty()) {
+      mine.setMinFloatValue(value);
+      mine.setMaxFloatValue(value);
+    } else {
+      if (value < mine.getMinFloatValue()) { mine.setMinFloatValue(value); }
+      if (value > mine.getMaxFloatValue()) { mine.setMaxFloatValue(value); }
+    }
+    if (mine.getLevelsArrayAt(0) == 0) { KllHelper.compressWhileUpdatingSketch(mine); }
+    mine.incN();
+    mine.setLevelZeroSorted(false);
+    final int nextPos = mine.getLevelsArrayAt(0) - 1;
+    assert mine.getLevelsArrayAt(0) >= 0;
+    mine.setLevelsArrayAt(0, nextPos);
+    mine.setFloatItemsArrayAt(nextPos, value);
+  }
 
   /**
    * Compression algorithm used to merge higher levels.
@@ -67,7 +343,7 @@ class KllFloatsHelper {
    * @param random instance of java.util.Random
    * @return int array of: {numLevels, targetItemCount, currentItemCount)
    */
-  static int[] generalFloatsCompress(
+  private static int[] generalFloatsCompress(
       final int k,
       final int m,
       final int numLevelsIn,
@@ -155,69 +431,103 @@ class KllFloatsHelper {
     return new int[] {numLevels, targetItemCount, currentItemCount};
   }
 
-  static void mergeSortedFloatArrays(
-      final float[] bufA, final int startA, final int lenA,
-      final float[] bufB, final int startB, final int lenB,
-      final float[] bufC, final int startC) {
-    final int lenC = lenA + lenB;
-    final int limA = startA + lenA;
-    final int limB = startB + lenB;
-    final int limC = startC + lenC;
+  private static KllFloatsQuantileCalculator getFloatsQuantileCalculator(final KllSketch mine) {
+    final int[] myLevelsArr = mine.getLevelsArray();
+    final float[] myFloatItemsArr = mine.getFloatItemsArray();
+    if (!mine.isLevelZeroSorted()) {
+      Arrays.sort(myFloatItemsArr, myLevelsArr[0], myLevelsArr[1]);
+      mine.setLevelZeroSorted(true);
+    }
+    return new KllFloatsQuantileCalculator(myFloatItemsArr, myLevelsArr, mine.getNumLevels(), mine.getN());
+  }
 
-    int a = startA;
-    int b = startB;
-
-    for (int c = startC; c < limC; c++) {
-      if (a == limA) {
-        bufC[c] = bufB[b];
-        b++;
-      } else if (b == limB) {
-        bufC[c] = bufA[a];
-        a++;
-      } else if (bufA[a] < bufB[b]) {
-        bufC[c] = bufA[a];
-        a++;
+  private static void incrementFloatBucketsSortedLevel(
+      final KllSketch mine, final int fromIndex, final int toIndex,
+      final int weight, final float[] splitPoints, final double[] buckets) {
+    final float[] myFloatItemsArr = mine.getFloatItemsArray();
+    int i = fromIndex;
+    int j = 0;
+    while (i <  toIndex && j < splitPoints.length) {
+      if (myFloatItemsArr[i] < splitPoints[j]) {
+        buckets[j] += weight; // this sample goes into this bucket
+        i++; // move on to next sample and see whether it also goes into this bucket
       } else {
-        bufC[c] = bufB[b];
-        b++;
+        j++; // no more samples for this bucket
       }
     }
-    assert a == limA;
-    assert b == limB;
-  }
-
-  //This must be modified for validation
-  static void randomlyHalveDownFloats(final float[] buf, final int start, final int length, final Random random) {
-    assert isEven(length);
-    final int half_length = length / 2;
-    final int offset = random.nextInt(2);       // disable for validation
-    //final int offset = deterministicOffset(); // enable for validation
-    int j = start + offset;
-    for (int i = start; i < (start + half_length); i++) {
-      buf[i] = buf[j];
-      j += 2;
+    // now either i == toIndex (we are out of samples), or
+    // j == numSplitPoints (we are out of buckets, but there are more samples remaining)
+    // we only need to do something in the latter case
+    if (j == splitPoints.length) {
+      buckets[j] += weight * (toIndex - i);
     }
   }
 
-  //This must be modified for validation
-  static void randomlyHalveUpFloats(final float[] buf, final int start, final int length, final Random random) {
-    assert isEven(length);
-    final int half_length = length / 2;
-    final int offset = random.nextInt(2);       // disable for validation
-    //final int offset = deterministicOffset(); // enable for validation
-    int j = (start + length) - 1 - offset;
-    for (int i = (start + length) - 1; i >= (start + half_length); i--) {
-      buf[i] = buf[j];
-      j -= 2;
+  private static void incrementFloatBucketsUnsortedLevel(
+      final KllSketch mine, final int fromIndex, final int toIndex,
+      final int weight, final float[] splitPoints, final double[] buckets) {
+    final float[] myFloatItemsArr = mine.getFloatItemsArray();
+    for (int i = fromIndex; i < toIndex; i++) {
+      int j;
+      for (j = 0; j < splitPoints.length; j++) {
+        if (myFloatItemsArr[i] < splitPoints[j]) {
+          break;
+        }
+      }
+      buckets[j] += weight;
     }
+  }
+
+  private static void populateFloatWorkArrays(final KllSketch mine, final KllSketch other, final float[] workbuf,
+      final int[] worklevels, final int provisionalNumLevels) {
+    worklevels[0] = 0;
+    final int[] myLevelsArr = mine.getLevelsArray();
+    final int[] otherLevelsArr = other.getLevelsArray();
+    final float[] myFloatItemsArr = mine.getFloatItemsArray();
+    final float[] otherFloatItemsArr = other.getFloatItemsArray();
+
+    // Note: the level zero data from "other" was already inserted into "self"
+    final int selfPopZero = KllHelper.currentLevelSize(0, mine.getNumLevels(), myLevelsArr);
+    System.arraycopy( myFloatItemsArr, myLevelsArr[0], workbuf, worklevels[0], selfPopZero);
+    worklevels[1] = worklevels[0] + selfPopZero;
+
+    for (int lvl = 1; lvl < provisionalNumLevels; lvl++) {
+      final int selfPop = KllHelper.currentLevelSize(lvl, mine.getNumLevels(), myLevelsArr);
+      final int otherPop = KllHelper.currentLevelSize(lvl, other.getNumLevels(), otherLevelsArr);
+      worklevels[lvl + 1] = worklevels[lvl] + selfPop + otherPop;
+
+      if (selfPop > 0 && otherPop == 0) {
+        System.arraycopy( myFloatItemsArr, myLevelsArr[lvl], workbuf, worklevels[lvl], selfPop);
+      } else if (selfPop == 0 && otherPop > 0) {
+        System.arraycopy(otherFloatItemsArr, otherLevelsArr[lvl], workbuf, worklevels[lvl], otherPop);
+      } else if (selfPop > 0 && otherPop > 0) {
+        mergeSortedFloatArrays( myFloatItemsArr, myLevelsArr[lvl], selfPop, otherFloatItemsArr,
+            otherLevelsArr[lvl], otherPop, workbuf, worklevels[lvl]);
+      }
+    }
+  }
+
+  private static float resolveFloatMaxValue(final float myMax, final float otherMax) {
+    if (Float.isNaN(myMax) && Float.isNaN(otherMax)) { return Float.NaN; }
+    if (Float.isNaN(myMax)) { return otherMax; }
+    if (Float.isNaN(otherMax)) { return myMax; }
+    return max(myMax, otherMax);
+  }
+
+  private static float resolveFloatMinValue(final float myMin, final float otherMin) {
+    if (Float.isNaN(myMin) && Float.isNaN(otherMin)) { return Float.NaN; }
+    if (Float.isNaN(myMin)) { return otherMin; }
+    if (Float.isNaN(otherMin)) { return myMin; }
+    return min(myMin, otherMin);
   }
 
   /**
+   * Validation Method.
    * Checks the sequential validity of the given array of float values.
    * They must be unique, monotonically increasing and not NaN.
    * @param values the given array of values
    */
-  static void validateFloatValues(final float[] values) {
+  private static void validateFloatValues(final float[] values) {
     for (int i = 0; i < values.length; i++) {
       if (!Float.isFinite(values[i])) {
         throw new SketchesArgumentException("Values must be finite");
@@ -230,11 +540,11 @@ class KllFloatsHelper {
   }
 
   /*
+   * Validation Method.
    * The following must be enabled for use with the KllFloatsValidationTest,
    * which is only enabled for manual testing. In addition, two methods
    * above need to be modified as commented.
    */
-
   //  static int nextOffset = 0;
   //
   //  private static int deterministicOffset() {
