+++ src/main/java/org/apache/datasketches/kll/KllHelper.java
@@ -19,29 +19,55 @@
 
 package org.apache.datasketches.kll;
 
+import static java.lang.Math.abs;
+import static java.lang.Math.ceil;
+import static java.lang.Math.exp;
+import static java.lang.Math.log;
+import static java.lang.Math.max;
+import static java.lang.Math.min;
 import static java.lang.Math.pow;
+import static java.lang.Math.round;
 import static org.apache.datasketches.Util.floorPowerOf2;
+import static org.apache.datasketches.Util.isOdd;
 import static org.apache.datasketches.kll.KllPreambleUtil.DATA_START_ADR;
-import static org.apache.datasketches.kll.KllSketch.CDF_COEF;
-import static org.apache.datasketches.kll.KllSketch.CDF_EXP;
-import static org.apache.datasketches.kll.KllSketch.PMF_COEF;
-import static org.apache.datasketches.kll.KllSketch.PMF_EXP;
+import static org.apache.datasketches.kll.KllPreambleUtil.DATA_START_ADR_SINGLE_ITEM;
+import static org.apache.datasketches.kll.KllPreambleUtil.PREAMBLE_INTS_EMPTY_SINGLE;
+import static org.apache.datasketches.kll.KllPreambleUtil.PREAMBLE_INTS_FULL;
+import static org.apache.datasketches.kll.KllPreambleUtil.SERIAL_VERSION_EMPTY_FULL;
+import static org.apache.datasketches.kll.KllPreambleUtil.SERIAL_VERSION_SINGLE;
+import static org.apache.datasketches.kll.KllPreambleUtil.SERIAL_VERSION_UPDATABLE;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryDoubleSketchFlag;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryEmptyFlag;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryFamilyID;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryK;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryLevelZeroSortedFlag;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryM;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryMinK;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryN;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryNumLevels;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryPreInts;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemorySerVer;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemorySingleItemFlag;
+import static org.apache.datasketches.kll.KllPreambleUtil.setMemoryUpdatableFlag;
 import static org.apache.datasketches.kll.KllSketch.SketchType.DOUBLES_SKETCH;
 
+import java.util.Arrays;
+
+import org.apache.datasketches.Family;
 import org.apache.datasketches.SketchesArgumentException;
+import org.apache.datasketches.Util;
 import org.apache.datasketches.kll.KllSketch.SketchType;
+import org.apache.datasketches.memory.WritableMemory;
 
 /**
- * This class provides some useful sketch analysis tools that are used internally and also can be used by
- * interested users to understand the internal structure of the sketch as well as the growth properties of the
- * sketch given a stream length.
+ * This class provides some useful sketch analysis tools that are used internally.
  *
  * @author lrhodes
  *
  */
-public class KllHelper {
+final class KllHelper {
 
-  public static class GrowthStats {
+  static class GrowthStats {
     SketchType sketchType;
     int k;
     int m;
@@ -53,7 +79,7 @@ public class KllHelper {
     int updatableBytes;
   }
 
-  public static class LevelStats {
+  static class LevelStats {
     long n;
     int numLevels;
     int items;
@@ -65,15 +91,273 @@ public class KllHelper {
     }
   }
 
+  static final double EPS_DELTA_THRESHOLD = 1E-6;
+  static final double MIN_EPS = 4.7634E-5;
+  static final double PMF_COEF = 2.446;
+  static final double PMF_EXP = 0.9433;
+  static final double CDF_COEF = 2.296;
+  static final double CDF_EXP = 0.9723;
+
   /**
    * This is the exact powers of 3 from 3^0 to 3^30 where the exponent is the index
    */
-  private static final long[] powersOfThree =
+  private static long[] powersOfThree =
       new long[] {1, 3, 9, 27, 81, 243, 729, 2187, 6561, 19683, 59049, 177147, 531441,
   1594323, 4782969, 14348907, 43046721, 129140163, 387420489, 1162261467,
   3486784401L, 10460353203L, 31381059609L, 94143178827L, 282429536481L,
   847288609443L, 2541865828329L, 7625597484987L, 22876792454961L, 68630377364883L,
   205891132094649L};
+
+  static void buildHeapKllSketchFromMemory(final KllSketch mine, final KllMemoryValidate memVal) {
+    final boolean doubleType = (mine.sketchType == DOUBLES_SKETCH);
+    mine.setLevelZeroSorted(memVal.level0Sorted);
+    mine.setN(memVal.n);
+    mine.setMinK(memVal.minK);
+    mine.setNumLevels(memVal.numLevels);
+    final int[] myLevelsArr = new int[mine.getNumLevels() + 1];
+
+    if (memVal.updatableMemory) {
+      memVal.levelsArrUpdatable.getIntArray(0, myLevelsArr, 0, mine.getNumLevels() + 1);
+      mine.setLevelsArray(myLevelsArr);
+      if (doubleType) {
+        mine.setMinDoubleValue(memVal.minMaxArrUpdatable.getDouble(0));
+        mine.setMaxDoubleValue(memVal.minMaxArrUpdatable.getDouble(Double.BYTES));
+        final int itemsCap = (int)memVal.itemsArrUpdatable.getCapacity() / Double.BYTES;
+        final double[] myItemsArr = new double[itemsCap];
+        memVal.itemsArrUpdatable.getDoubleArray(0, myItemsArr, 0, itemsCap);
+        mine.setDoubleItemsArray(myItemsArr);
+      } else { //float
+        mine.setMinFloatValue(memVal.minMaxArrUpdatable.getFloat(0));
+        mine.setMaxFloatValue(memVal.minMaxArrUpdatable.getFloat(Float.BYTES));
+        final int itemsCap = (int)memVal.itemsArrUpdatable.getCapacity() / Float.BYTES;
+        final float[] myItemsArr = new float[itemsCap];
+        memVal.itemsArrUpdatable.getFloatArray(0, myItemsArr, 0, itemsCap);
+        mine.setFloatItemsArray(myItemsArr);
+      }
+    } else { //compact
+      memVal.levelsArrCompact.getIntArray(0, myLevelsArr, 0, mine.getNumLevels() + 1);
+      mine.setLevelsArray(myLevelsArr);
+      if (doubleType) {
+        mine.setMinDoubleValue(memVal.minMaxArrCompact.getDouble(0));
+        mine.setMaxDoubleValue(memVal.minMaxArrCompact.getDouble(Double.BYTES));
+        final int itemsCap = (int)memVal.itemsArrCompact.getCapacity() / Double.BYTES;
+        final double[] myItemsArr = new double[itemsCap];
+        memVal.itemsArrCompact.getDoubleArray(0, myItemsArr, 0, itemsCap);
+        mine.setDoubleItemsArray(myItemsArr);
+      } else { //float
+        mine.setMinFloatValue(memVal.minMaxArrCompact.getFloat(0));
+        mine.setMaxFloatValue(memVal.minMaxArrCompact.getFloat(Float.BYTES));
+        final int itemsCap = (int)memVal.itemsArrCompact.getCapacity() / Float.BYTES;
+        final float[] myItemsArr = new float[itemsCap];
+        memVal.itemsArrCompact.getFloatArray(0, myItemsArr, 0, itemsCap);
+        mine.setFloatItemsArray(myItemsArr);
+      }
+    }
+  }
+
+  /**
+   * Checks the validity of the given value k
+   * @param k must be greater than 7 and less than 65536.
+   */
+  static void checkK(final int k, final int m) {
+    if (k < m || k > KllSketch.MAX_K) {
+      throw new SketchesArgumentException(
+          "K must be >= " + m + " and <= " + KllSketch.MAX_K + ": " + k);
+    }
+  }
+
+  static void checkM(final int m) {
+    if (m < KllSketch.MIN_M || m > KllSketch.MAX_M || ((m & 1) == 1)) {
+      throw new SketchesArgumentException(
+          "M must be >= 2, <= 8 and even: " + m);
+    }
+  }
+
+  /**
+   * The following code is only valid in the special case of exactly reaching capacity while updating.
+   * It cannot be used while merging, while reducing k, or anything else.
+   * @param mine the current sketch
+   */
+  static void compressWhileUpdatingSketch(final KllSketch mine) {
+    final int level =
+        findLevelToCompact(mine.getK(), mine.getM(), mine.getNumLevels(), mine.getLevelsArray());
+    if (level == mine.getNumLevels() - 1) {
+      //The level to compact is the top level, thus we need to add a level.
+      //Be aware that this operation grows the items array,
+      //shifts the items data and the level boundaries of the data,
+      //and grows the levels array and increments numLevels_.
+      KllHelper.addEmptyTopLevelToCompletelyFullSketch(mine);
+    }
+
+    final int[] myLevelsArr = mine.getLevelsArray();
+    final int rawBeg = myLevelsArr[level];
+    final int rawEnd = myLevelsArr[level + 1];
+    // +2 is OK because we already added a new top level if necessary
+    final int popAbove = myLevelsArr[level + 2] - rawEnd;
+    final int rawPop = rawEnd - rawBeg;
+    final boolean oddPop = isOdd(rawPop);
+    final int adjBeg = oddPop ? rawBeg + 1 : rawBeg;
+    final int adjPop = oddPop ? rawPop - 1 : rawPop;
+    final int halfAdjPop = adjPop / 2;
+
+    // level zero might not be sorted, so we must sort it if we wish to compact it
+    float[] myFloatItemsArr;
+    double[] myDoubleItemsArr;
+
+    if (mine.sketchType == DOUBLES_SKETCH) {
+      myFloatItemsArr = null;
+      myDoubleItemsArr = mine.getDoubleItemsArray();
+      if (level == 0) {
+        if (mine.updatablMemory) {
+          myDoubleItemsArr = mine.getDoubleItemsArray();
+          Arrays.sort(myDoubleItemsArr, adjBeg, adjBeg + adjPop);
+          mine.setDoubleItemsArray(myDoubleItemsArr);
+        } else {
+          Arrays.sort(mine.getDoubleItemsArray(), adjBeg, adjBeg + adjPop);
+        }
+      }
+      if (popAbove == 0) {
+        if (mine.updatablMemory) {
+          myDoubleItemsArr = mine.getDoubleItemsArray();
+          KllDoublesHelper.randomlyHalveUpDoubles(myDoubleItemsArr, adjBeg, adjPop, KllSketch.random);
+          mine.setDoubleItemsArray(myDoubleItemsArr);
+        } else {
+          KllDoublesHelper.randomlyHalveUpDoubles(mine.getDoubleItemsArray(), adjBeg, adjPop, KllSketch.random);
+        }
+      } else {
+        if (mine.updatablMemory) {
+          myDoubleItemsArr = mine.getDoubleItemsArray();
+          KllDoublesHelper.randomlyHalveDownDoubles(myDoubleItemsArr, adjBeg, adjPop, KllSketch.random);
+          mine.setDoubleItemsArray(myDoubleItemsArr);
+        } else {
+          KllDoublesHelper.randomlyHalveDownDoubles(mine.getDoubleItemsArray(), adjBeg, adjPop, KllSketch.random);
+        }
+        if (mine.updatablMemory ) {
+          myDoubleItemsArr = mine.getDoubleItemsArray();
+          KllDoublesHelper.mergeSortedDoubleArrays(
+              myDoubleItemsArr, adjBeg, halfAdjPop,
+              myDoubleItemsArr, rawEnd, popAbove,
+              myDoubleItemsArr, adjBeg + halfAdjPop);
+          mine.setDoubleItemsArray(myDoubleItemsArr);
+        } else {
+          myDoubleItemsArr = mine.getDoubleItemsArray();
+          KllDoublesHelper.mergeSortedDoubleArrays(
+              myDoubleItemsArr, adjBeg, halfAdjPop,
+              myDoubleItemsArr, rawEnd, popAbove,
+              myDoubleItemsArr, adjBeg + halfAdjPop);
+        }
+      }
+    } else { //Float sketch
+      myFloatItemsArr = mine.getFloatItemsArray();
+      myDoubleItemsArr = null;
+      if (level == 0) {
+        if (mine.updatablMemory) {
+          myFloatItemsArr = mine.getFloatItemsArray();
+          Arrays.sort(myFloatItemsArr, adjBeg, adjBeg + adjPop);
+          mine.setFloatItemsArray(myFloatItemsArr);
+        } else {
+          Arrays.sort(mine.getFloatItemsArray(), adjBeg, adjBeg + adjPop);
+        }
+      }
+      if (popAbove == 0) {
+        if (mine.updatablMemory) {
+          myFloatItemsArr = mine.getFloatItemsArray();
+          KllFloatsHelper.randomlyHalveUpFloats(myFloatItemsArr, adjBeg, adjPop, KllSketch.random);
+          mine.setFloatItemsArray(myFloatItemsArr);
+        } else {
+          KllFloatsHelper.randomlyHalveUpFloats(mine.getFloatItemsArray(), adjBeg, adjPop, KllSketch.random);
+        }
+      } else {
+        if (mine.updatablMemory) {
+          myFloatItemsArr = mine.getFloatItemsArray();
+          KllFloatsHelper.randomlyHalveDownFloats(myFloatItemsArr, adjBeg, adjPop, KllSketch.random);
+          mine.setFloatItemsArray(myFloatItemsArr);
+        } else {
+          KllFloatsHelper.randomlyHalveDownFloats(mine.getFloatItemsArray(), adjBeg, adjPop, KllSketch.random);
+        }
+        if (mine.updatablMemory ) {
+          myFloatItemsArr = mine.getFloatItemsArray();
+          KllFloatsHelper.mergeSortedFloatArrays(
+              myFloatItemsArr, adjBeg, halfAdjPop,
+              myFloatItemsArr, rawEnd, popAbove,
+              myFloatItemsArr, adjBeg + halfAdjPop);
+          mine.setFloatItemsArray(myFloatItemsArr);
+        } else {
+          myFloatItemsArr = mine.getFloatItemsArray();
+          KllFloatsHelper.mergeSortedFloatArrays(
+              myFloatItemsArr, adjBeg, halfAdjPop,
+              myFloatItemsArr, rawEnd, popAbove,
+              myFloatItemsArr, adjBeg + halfAdjPop);
+        }
+      }
+    }
+    mine.setLevelsArrayAtMinusEq(level + 1, halfAdjPop); // adjust boundaries of the level above
+
+    if (oddPop) {
+      mine.setLevelsArrayAt(level, mine.getLevelsArrayAt(level + 1) - 1); // the current level now contains one item
+      if (mine.sketchType == DOUBLES_SKETCH) {
+        mine.setDoubleItemsArrayAt(
+            mine.getLevelsArrayAt(level), mine.getDoubleItemsArrayAt(rawBeg)); // namely this leftover guy
+      } else {
+        mine.setFloatItemsArrayAt(
+            mine.getLevelsArrayAt(level), mine.getFloatItemsArrayAt(rawBeg)); // namely this leftover guy
+      }
+
+    } else {
+      mine.setLevelsArrayAt(level, mine.getLevelsArrayAt(level + 1)); // the current level is now empty
+    }
+
+    // verify that we freed up halfAdjPop array slots just below the current level
+    assert mine.getLevelsArrayAt(level) == rawBeg + halfAdjPop;
+
+    // finally, we need to shift up the data in the levels below
+    // so that the freed-up space can be used by level zero
+    if (level > 0) {
+      final int amount = rawBeg - mine.getLevelsArrayAt(0);
+      if (mine.sketchType == DOUBLES_SKETCH) {
+        if (mine.updatablMemory) {
+          myDoubleItemsArr = mine.getDoubleItemsArray();
+          System.arraycopy(myDoubleItemsArr, myLevelsArr[0], myDoubleItemsArr, myLevelsArr[0] + halfAdjPop, amount);
+          mine.setDoubleItemsArray(myDoubleItemsArr);
+        } else {
+          System.arraycopy(myDoubleItemsArr, myLevelsArr[0], myDoubleItemsArr, myLevelsArr[0] + halfAdjPop, amount);
+        }
+      } else {
+        if (mine.updatablMemory) {
+          myFloatItemsArr = mine.getFloatItemsArray();
+          System.arraycopy(myFloatItemsArr, myLevelsArr[0], myFloatItemsArr, myLevelsArr[0] + halfAdjPop, amount);
+          mine.setFloatItemsArray(myFloatItemsArr);
+        } else {
+          System.arraycopy(myFloatItemsArr, myLevelsArr[0], myFloatItemsArr, myLevelsArr[0] + halfAdjPop, amount);
+        }
+      }
+      for (int lvl = 0; lvl < level; lvl++) {
+        mine.setLevelsArrayAtPlusEq(lvl, halfAdjPop);
+      }
+    }
+  }
+
+  /**
+   * Returns the maximum number of items that this sketch can handle
+   * @param k The sizing / accuracy parameter of the sketch in items.
+   * Note: this method actually works for k values up to k = 2^29 and 61 levels,
+   * however only k values up to (2^16 - 1) are currently used by the sketch.
+   * @param m the size of the smallest level in items. Default is 8.
+   * @param numLevels the upper bound number of levels based on <i>n</i> items.
+   * @return the total item capacity of the sketch.
+   */
+  static int computeTotalItemCapacity(final int k, final int m, final int numLevels) {
+    long total = 0;
+    for (int level = 0; level < numLevels; level++) {
+      total += levelCapacity(k, numLevels, level, m);
+    }
+    return (int) total;
+  }
+
+  static int currentLevelSize(final int level, final int numLevels, final int[] levels) {
+    if (level >= numLevels) { return 0; }
+    return levels[level + 1] - levels[level];
+  }
 
   /**
    * Given k, m, and numLevels, this computes and optionally prints the structure of the sketch when the given
@@ -85,7 +369,7 @@ public class KllHelper {
    * @return LevelStats with the final summary of the sketch's cumulative N,
    * and cumulative items at the given numLevels.
    */
-  public static LevelStats getFinalSketchStatsAtNumLevels(
+  static LevelStats getFinalSketchStatsAtNumLevels(
       final int k,
       final int m,
       final int numLevels,
@@ -120,7 +404,7 @@ public class KllHelper {
    * @param printGrowthScheme if true the entire growth scheme of the sketch will be printed.
    * @return GrowthStats with the final values of the growth scheme
    */
-  public static GrowthStats getGrowthSchemeForGivenN(
+  static GrowthStats getGrowthSchemeForGivenN(
       final int k,
       final int m,
       final long n,
@@ -163,6 +447,20 @@ public class KllHelper {
     return gStats;
   }
 
+  // constants were derived as the best fit to 99 percentile empirically measured max error in
+  // thousands of trials
+  static int getKFromEpsilon(final double epsilon, final boolean pmf) {
+    //Ensure that eps is >= than the lowest possible eps given MAX_K and pmf=false.
+    final double eps = max(epsilon, MIN_EPS);
+    final double kdbl = pmf
+        ? exp(log(PMF_COEF / eps) / PMF_EXP)
+        : exp(log(CDF_COEF / eps) / CDF_EXP);
+    final double krnd = round(kdbl);
+    final double del = abs(krnd - kdbl);
+    final int k = (int) (del < EPS_DELTA_THRESHOLD ? krnd : ceil(kdbl));
+    return max(KllSketch.MIN_M, min(KllSketch.MAX_K, k));
+  }
+
   /**
    * Given k, m, numLevels, this computes the item capacity of a single level.
    * @param k the given user sketch configuration parameter
@@ -171,7 +469,7 @@ public class KllHelper {
    * @param level the specific level to compute its item capacity
    * @return LevelStats with the computed N and items for the given level.
    */
-  public static LevelStats getLevelCapacityItems(
+  static LevelStats getLevelCapacityItems(
       final int k,
       final int m,
       final int numLevels,
@@ -179,66 +477,6 @@ public class KllHelper {
     final int items = KllHelper.levelCapacity(k, numLevels, level, m);
     final long n = (long)items << level;
     return new LevelStats(n, numLevels, items);
-  }
-
-  /**
-   * Checks the validity of the given value k
-   * @param k must be greater than 7 and less than 65536.
-   */
-  static void checkK(final int k, final int m) {
-    if (k < m || k > KllSketch.MAX_K) {
-      throw new SketchesArgumentException(
-          "K must be >= " + m + " and <= " + KllSketch.MAX_K + ": " + k);
-    }
-  }
-
-  static void checkM(final int m) {
-    if (m < KllSketch.MIN_M || m > KllSketch.MAX_M || ((m & 1) == 1)) {
-      throw new SketchesArgumentException(
-          "M must be >= 2, <= 8 and even: " + m);
-    }
-  }
-
-  /**
-   * Returns the maximum number of items that this sketch can handle
-   * @param k The sizing / accuracy parameter of the sketch in items.
-   * Note: this method actually works for k values up to k = 2^29 and 61 levels,
-   * however only k values up to (2^16 - 1) are currently used by the sketch.
-   * @param m the size of the smallest level in items. Default is 8.
-   * @param numLevels the upper bound number of levels based on <i>n</i> items.
-   * @return the total item capacity of the sketch.
-   */
-  static int computeTotalItemCapacity(final int k, final int m, final int numLevels) {
-    long total = 0;
-    for (int level = 0; level < numLevels; level++) {
-      total += levelCapacity(k, numLevels, level, m);
-    }
-    return (int) total;
-  }
-
-  static int currentLevelSize(final int level, final int numLevels, final int[] levels) {
-    if (level >= numLevels) { return 0; }
-    return levels[level + 1] - levels[level];
-  }
-
-  /**
-   * Finds the first level starting with level 0 that exceeds its nominal capacity
-   * @param k configured size of sketch. Range [m, 2^16]
-   * @param m minimum level size. Default is 8.
-   * @param numLevels one-based number of current levels
-   * @return level to compact
-   */
-  static int findLevelToCompact(final int k, final int m, final int numLevels, final int[] levels) {
-    int level = 0;
-    while (true) {
-      assert level < numLevels;
-      final int pop = levels[level + 1] - levels[level];
-      final int cap = KllHelper.levelCapacity(k, numLevels, level, m);
-      if (pop >= cap) {
-        return level;
-      }
-      level++;
-    }
   }
 
   /**
@@ -281,6 +519,131 @@ public class KllHelper {
     return (int) Math.max(m, intCapAux(k, depth));
   }
 
+  /**
+   * This method is for direct Double and Float sketches only and does the following:
+   * <ul>
+   * <li>Determines if the required sketch bytes will fit in the current Memory.
+   * If so, it will stretch the positioning of the arrays to fit. Otherwise:
+   * <li>Allocates a new WritableMemory of the required size</li>
+   * <li>Copies over the preamble as is (20 bytes)</li>
+   * <li>Creates new memory regions for Levels Array, Min/Max Array, Items Array, but
+   * does not fill them. They may contain garbage.</li>
+   * </ul>
+   * The caller is responsible for filling these regions and updating the preamble.
+   * @param sketch The current sketch that needs to be expanded.
+   * @param newLevelsArrLen the element length of the new Levels array.
+   * @param newItemsArrLen the element length of the new Items array.
+   * @return the new expanded memory with preamble.
+   */
+  static WritableMemory memorySpaceMgmt(
+      final KllSketch sketch,
+      final int newLevelsArrLen,
+      final int newItemsArrLen) {
+    final KllSketch.SketchType sketchType = sketch.sketchType;
+    final WritableMemory oldWmem = sketch.wmem;
+    final int startAdr = DATA_START_ADR;
+    final int typeBytes = (sketchType == DOUBLES_SKETCH) ? Double.BYTES : Float.BYTES;
+
+    int requiredSketchBytes = startAdr;
+    requiredSketchBytes += newLevelsArrLen * Integer.BYTES;
+    requiredSketchBytes += 2 * typeBytes;
+    requiredSketchBytes += newItemsArrLen * typeBytes;
+    final WritableMemory newWmem;
+
+    if (requiredSketchBytes > oldWmem.getCapacity()) { //Acquire new WritableMemory
+      newWmem = sketch.memReqSvr.request(oldWmem, requiredSketchBytes);
+      oldWmem.copyTo(0, newWmem, 0, startAdr); //copy preamble
+    }
+    else { //Expand or contract in current memory
+      newWmem = oldWmem;
+    }
+
+    int offset = startAdr;
+    //LEVELS ARR
+    int lengthBytes = newLevelsArrLen * Integer.BYTES;
+    sketch.setLevelsArrayUpdatable(newWmem.writableRegion(offset, lengthBytes)); //
+    offset += lengthBytes;
+    //MIN MAX ARR
+    lengthBytes = 2 * typeBytes;
+    sketch.setMinMaxArrayUpdatable(newWmem.writableRegion(offset, lengthBytes));
+    offset += lengthBytes;
+    //ITEMS ARR
+    lengthBytes = newItemsArrLen * typeBytes;
+    sketch.setItemsArrayUpdatable(newWmem.writableRegion(offset, lengthBytes));
+    assert requiredSketchBytes <= newWmem.getCapacity();
+    return newWmem;
+  }
+
+  static String outputData(final boolean doubleType, final int numLevels, final int[] levelsArr,
+      final float[] floatItemsArr, final double[] doubleItemsArr) {
+    final StringBuilder sb =  new StringBuilder();
+    sb.append("### KLL items data {index, item}:").append(Util.LS);
+    if (levelsArr[0] > 0) {
+      sb.append(" Garbage:" + Util.LS);
+      if (doubleType) {
+        for (int i = 0; i < levelsArr[0]; i++) {
+          sb.append("   ").append(i + ", ").append(doubleItemsArr[i]).append(Util.LS);
+        }
+      } else {
+        for (int i = 0; i < levelsArr[0]; i++) {
+          sb.append("   ").append(i + ", ").append(floatItemsArr[i]).append(Util.LS);
+        }
+      }
+    }
+    int level = 0;
+    if (doubleType) {
+      while (level < numLevels) {
+        final int fromIndex = levelsArr[level];
+        final int toIndex = levelsArr[level + 1]; // exclusive
+        if (fromIndex < toIndex) {
+          sb.append(" level[").append(level).append("]: offset: " + levelsArr[level] + " wt: " + (1 << level));
+          sb.append(Util.LS);
+        }
+
+        for (int i = fromIndex; i < toIndex; i++) {
+          sb.append("   ").append(i + ", ").append(doubleItemsArr[i]).append(Util.LS);
+        }
+        level++;
+      }
+    }
+    else {
+      while (level < numLevels) {
+        final int fromIndex = levelsArr[level];
+        final int toIndex = levelsArr[level + 1]; // exclusive
+        if (fromIndex <= toIndex) {
+          sb.append(" level[").append(level).append("]: offset: " + levelsArr[level] + " wt: " + (1 << level));
+          sb.append(Util.LS);
+        }
+
+        for (int i = fromIndex; i < toIndex; i++) {
+          sb.append("   ").append(i + ", ").append(floatItemsArr[i]).append(Util.LS);
+        }
+        level++;
+      }
+    }
+    sb.append(" level[" + level + "]: offset: " + levelsArr[level] + " (Exclusive)");
+    sb.append(Util.LS);
+    sb.append("### End items data").append(Util.LS);
+
+    return sb.toString();
+  }
+
+  static String outputLevels(final int k, final int m, final int numLevels, final int[] levelsArr) {
+    final StringBuilder sb =  new StringBuilder();
+    sb.append("### KLL levels array:").append(Util.LS)
+    .append(" level, offset: nominal capacity, actual size").append(Util.LS);
+    int level = 0;
+    for ( ; level < numLevels; level++) {
+      sb.append("   ").append(level).append(", ").append(levelsArr[level]).append(": ")
+      .append(KllHelper.levelCapacity(k, numLevels, level, m))
+      .append(", ").append(KllHelper.currentLevelSize(level, numLevels, levelsArr)).append(Util.LS);
+    }
+    sb.append("   ").append(level).append(", ").append(levelsArr[level]).append(": (Exclusive)")
+    .append(Util.LS);
+    sb.append("### End levels array").append(Util.LS);
+    return sb.toString();
+  }
+
   static long sumTheSampleWeights(final int num_levels, final int[] levels) {
     long total = 0;
     long weight = 1;
@@ -291,6 +654,146 @@ public class KllHelper {
     return total;
   }
 
+  static byte[] toCompactByteArrayImpl(final KllSketch mine) {
+    final byte[] byteArr = new byte[mine.getCurrentCompactSerializedSizeBytes()];
+    final WritableMemory wmem = WritableMemory.writableWrap(byteArr);
+    loadFirst8Bytes(mine, wmem, false);
+    if (mine.getN() == 0) { return byteArr; } //empty
+    final boolean doubleType = (mine.sketchType == DOUBLES_SKETCH);
+
+    //load data
+    int offset = DATA_START_ADR_SINGLE_ITEM;
+    final int[] myLevelsArr = mine.getLevelsArray();
+    if (mine.getN() == 1) { //single item
+      if (doubleType) {
+        wmem.putDouble(offset,  mine.getDoubleItemsArray()[myLevelsArr[0]]);
+      } else {
+        wmem.putFloat(offset, mine.getFloatItemsArray()[myLevelsArr[0]]);
+      }
+    } else { // n > 1
+      //remainder of preamble after first 8 bytes
+      setMemoryN(wmem, mine.getN());
+      setMemoryMinK(wmem, mine.getMinK());
+      setMemoryNumLevels(wmem, mine.getNumLevels());
+      offset = DATA_START_ADR;
+
+      //LOAD LEVELS ARR the last integer in levels_ is NOT serialized
+      final int len = myLevelsArr.length - 1;
+      wmem.putIntArray(offset, myLevelsArr, 0, len);
+      offset += len * Integer.BYTES;
+
+      //LOAD MIN, MAX VALUES FOLLOWED BY ITEMS ARRAY
+      if (doubleType) {
+        wmem.putDouble(offset,mine. getMinDoubleValue());
+        offset += Double.BYTES;
+        wmem.putDouble(offset, mine.getMaxDoubleValue());
+        offset += Double.BYTES;
+        wmem.putDoubleArray(offset, mine.getDoubleItemsArray(), myLevelsArr[0], mine.getNumRetained());
+      } else {
+        wmem.putFloat(offset, mine.getMinFloatValue());
+        offset += Float.BYTES;
+        wmem.putFloat(offset, mine.getMaxFloatValue());
+        offset += Float.BYTES;
+        wmem.putFloatArray(offset, mine.getFloatItemsArray(), myLevelsArr[0], mine.getNumRetained());
+      }
+    }
+    return byteArr;
+  }
+
+  @SuppressWarnings("null")
+  static String toStringImpl(final KllSketch mine, final boolean withLevels, final boolean withData) {
+    final boolean doubleType = (mine.sketchType == DOUBLES_SKETCH);
+    final int k = mine.getK();
+    final int m = mine.getM();
+    final String epsPct = String.format("%.3f%%", mine.getNormalizedRankError(false) * 100);
+    final String epsPMFPct = String.format("%.3f%%", mine.getNormalizedRankError(true) * 100);
+    final StringBuilder sb = new StringBuilder();
+    final String skType = (mine.updatablMemory ? "Direct" : "") + (doubleType ? "Doubles" : "Floats");
+    sb.append(Util.LS).append("### Kll").append(skType).append("Sketch Summary:").append(Util.LS);
+    sb.append("   K                      : ").append(k).append(Util.LS);
+    sb.append("   Dynamic min K          : ").append(mine.getMinK()).append(Util.LS);
+    sb.append("   M                      : ").append(m).append(Util.LS);
+    sb.append("   N                      : ").append(mine.getN()).append(Util.LS);
+    sb.append("   Epsilon                : ").append(epsPct).append(Util.LS);
+    sb.append("   Epsison PMF            : ").append(epsPMFPct).append(Util.LS);
+    sb.append("   Empty                  : ").append(mine.isEmpty()).append(Util.LS);
+    sb.append("   Estimation Mode        : ").append(mine.isEstimationMode()).append(Util.LS);
+    sb.append("   Levels                 : ").append(mine.getNumLevels()).append(Util.LS);
+    sb.append("   Level 0 Sorted         : ").append(mine.isLevelZeroSorted()).append(Util.LS);
+    final int cap = (doubleType) ? mine.getDoubleItemsArray().length : mine.getFloatItemsArray().length;
+    sb.append("   Capacity Items         : ").append(cap).append(Util.LS);
+    sb.append("   Retained Items         : ").append(mine.getNumRetained()).append(Util.LS);
+    if (mine.updatablMemory) {
+      sb.append("   Updatable Storage Bytes: ").append(mine.getCurrentUpdatableSerializedSizeBytes()).append(Util.LS);
+    } else {
+      sb.append("   Compact Storage Bytes  : ").append(mine.getCurrentCompactSerializedSizeBytes()).append(Util.LS);
+    }
+
+    if (doubleType) {
+      sb.append("   Min Value              : ").append(mine.getMinDoubleValue()).append(Util.LS);
+      sb.append("   Max Value              : ").append(mine.getMaxDoubleValue()).append(Util.LS);
+    } else {
+      sb.append("   Min Value              : ").append(mine.getMinFloatValue()).append(Util.LS);
+      sb.append("   Max Value              : ").append(mine.getMaxFloatValue()).append(Util.LS);
+    }
+    sb.append("### End sketch summary").append(Util.LS);
+
+    final int myNumLevels = mine.getNumLevels();
+    final int[] myLevelsArr = mine.getLevelsArray();
+    double[] myDoubleItemsArr = null;
+    float[] myFloatItemsArr = null;
+    if (doubleType) {
+      myDoubleItemsArr = mine.getDoubleItemsArray();
+    } else {
+      myFloatItemsArr = mine.getFloatItemsArray();
+    }
+    if (withLevels) {
+      sb.append(outputLevels(k, m, myNumLevels, myLevelsArr));
+    }
+    if (withData) {
+      sb.append(outputData(doubleType, myNumLevels, myLevelsArr, myFloatItemsArr, myDoubleItemsArr));
+    }
+    return sb.toString();
+  }
+
+  static byte[] toUpdatableByteArrayImpl(final KllSketch mine) {
+    final byte[] byteArr = new byte[mine.getCurrentUpdatableSerializedSizeBytes()];
+    final WritableMemory wmem = WritableMemory.writableWrap(byteArr);
+    loadFirst8Bytes(mine, wmem, true);
+    //remainder of preamble after first 8 bytes
+    setMemoryN(wmem, mine.getN());
+    setMemoryMinK(wmem, mine.getMinK());
+    setMemoryNumLevels(wmem, mine.getNumLevels());
+
+    //load data
+    final boolean doubleType = (mine.sketchType == DOUBLES_SKETCH);
+    int offset = DATA_START_ADR;
+
+    //LOAD LEVELS ARRAY the last integer in levels_ IS serialized
+    final int[] myLevelsArr = mine.getLevelsArray();
+    final int len = myLevelsArr.length;
+    wmem.putIntArray(offset, myLevelsArr, 0, len);
+    offset += len * Integer.BYTES;
+
+    //LOAD MIN, MAX VALUES FOLLOWED BY ITEMS ARRAY
+    if (doubleType) {
+      wmem.putDouble(offset, mine.getMinDoubleValue());
+      offset += Double.BYTES;
+      wmem.putDouble(offset, mine.getMaxDoubleValue());
+      offset += Double.BYTES;
+      final double[] doubleItemsArr = mine.getDoubleItemsArray();
+      wmem.putDoubleArray(offset, doubleItemsArr, 0, doubleItemsArr.length);
+    } else {
+      wmem.putFloat(offset, mine.getMinFloatValue());
+      offset += Float.BYTES;
+      wmem.putFloat(offset,mine.getMaxFloatValue());
+      offset += Float.BYTES;
+      final float[] floatItemsArr = mine.getFloatItemsArray();
+      wmem.putFloatArray(offset, floatItemsArr, 0, floatItemsArr.length);
+    }
+    return byteArr;
+  }
+
   /**
    * Returns very conservative upper bound of the number of levels based on <i>n</i>.
    * @param n the length of the stream
@@ -298,6 +801,118 @@ public class KllHelper {
    */
   static int ubOnNumLevels(final long n) {
     return 1 + Long.numberOfTrailingZeros(floorPowerOf2(n));
+  }
+
+  /**
+   * This grows the levels arr by 1 (if needed) and increases the capacity of the items array
+   * at the bottom.  Only numLevels, the levels array and the items array are affected.
+   * @param mine the current sketch
+   */
+  @SuppressWarnings("null")
+  private static void addEmptyTopLevelToCompletelyFullSketch(final KllSketch mine) {
+    final int[] myCurLevelsArr = mine.getLevelsArray();
+    final int myCurNumLevels = mine.getNumLevels();
+    final int myCurTotalItemsCapacity = myCurLevelsArr[myCurNumLevels];
+    double minDouble = Double.NaN;
+    double maxDouble = Double.NaN;
+    float minFloat = Float.NaN;
+    float maxFloat = Float.NaN;
+
+    double[] myCurDoubleItemsArr = null;
+    float[] myCurFloatItemsArr = null;
+
+    final int myNewNumLevels;
+    final int[] myNewLevelsArr;
+    final int myNewTotalItemsCapacity;
+
+    float[] myNewFloatItemsArr = null;
+    double[] myNewDoubleItemsArr = null;
+
+    if (mine.sketchType == DOUBLES_SKETCH) {
+      minDouble = mine.getMinDoubleValue();
+      maxDouble = mine.getMaxDoubleValue();
+      myCurDoubleItemsArr = mine.getDoubleItemsArray();
+      //assert we are following a certain growth scheme
+      assert myCurDoubleItemsArr.length == myCurTotalItemsCapacity;
+    } else { //FLOATS_SKETCH
+      minFloat = mine.getMinFloatValue();
+      maxFloat = mine.getMaxFloatValue();
+      myCurFloatItemsArr = mine.getFloatItemsArray();
+      assert myCurFloatItemsArr.length == myCurTotalItemsCapacity;
+    }
+    assert myCurLevelsArr[0] == 0; //definition of full is part of the growth scheme
+
+    final int deltaItemsCap = levelCapacity(mine.getK(), myCurNumLevels + 1, 0, mine.getM());
+    myNewTotalItemsCapacity = myCurTotalItemsCapacity + deltaItemsCap;
+
+    // Check if growing the levels arr if required.
+    // Note that merging MIGHT over-grow levels_, in which case we might not have to grow it
+    final boolean growLevelsArr = myCurLevelsArr.length < myCurNumLevels + 2;
+
+    // GROW LEVELS ARRAY
+    if (growLevelsArr) {
+      //grow levels arr by one and copy the old data to the new array, extra space at the top.
+      myNewLevelsArr = Arrays.copyOf(myCurLevelsArr, myCurNumLevels + 2);
+      assert myNewLevelsArr.length == myCurLevelsArr.length + 1;
+      myNewNumLevels = myCurNumLevels + 1;
+      mine.incNumLevels(); //increment the class member
+    } else {
+      myNewLevelsArr = myCurLevelsArr;
+      myNewNumLevels = myCurNumLevels;
+    }
+    // This loop updates all level indices EXCLUDING the "extra" index at the top
+    for (int level = 0; level <= myNewNumLevels - 1; level++) {
+      myNewLevelsArr[level] += deltaItemsCap;
+    }
+    myNewLevelsArr[myNewNumLevels] = myNewTotalItemsCapacity; // initialize the new "extra" index at the top
+
+    // GROW ITEMS ARRAY
+    if (mine.sketchType == DOUBLES_SKETCH) {
+      myNewDoubleItemsArr = new double[myNewTotalItemsCapacity];
+      // copy and shift the current data into the new array
+      System.arraycopy(myCurDoubleItemsArr, 0, myNewDoubleItemsArr, deltaItemsCap, myCurTotalItemsCapacity);
+    } else {
+      myNewFloatItemsArr = new float[myNewTotalItemsCapacity];
+      // copy and shift the current items data into the new array
+      System.arraycopy(myCurFloatItemsArr, 0, myNewFloatItemsArr, deltaItemsCap, myCurTotalItemsCapacity);
+    }
+
+    //MEMORY SPACE MANAGEMENT
+    if (mine.updatablMemory) {
+      mine.wmem = memorySpaceMgmt(mine, myNewLevelsArr.length, myNewTotalItemsCapacity);
+    }
+    //update our sketch with new expanded spaces
+    mine.setNumLevels(myNewNumLevels);
+    mine.setLevelsArray(myNewLevelsArr);
+    if (mine.sketchType == DOUBLES_SKETCH) {
+      mine.setMinDoubleValue(minDouble);
+      mine.setMaxDoubleValue(maxDouble);
+      mine.setDoubleItemsArray(myNewDoubleItemsArr);
+    } else { //Float sketch
+      mine.setMinFloatValue(minFloat);
+      mine.setMaxFloatValue(maxFloat);
+      mine.setFloatItemsArray(myNewFloatItemsArr);
+    }
+  }
+
+  /**
+   * Finds the first level starting with level 0 that exceeds its nominal capacity
+   * @param k configured size of sketch. Range [m, 2^16]
+   * @param m minimum level size. Default is 8.
+   * @param numLevels one-based number of current levels
+   * @return level to compact
+   */
+  private static int findLevelToCompact(final int k, final int m, final int numLevels, final int[] levels) {
+    int level = 0;
+    while (true) {
+      assert level < numLevels;
+      final int pop = levels[level + 1] - levels[level];
+      final int cap = KllHelper.levelCapacity(k, numLevels, level, m);
+      if (pop >= cap) {
+        return level;
+      }
+      level++;
+    }
   }
 
   /**
@@ -331,6 +946,30 @@ public class KllHelper {
     return result;
   }
 
+  private static void loadFirst8Bytes(final KllSketch sk, final WritableMemory wmem,
+      final boolean updatable) {
+    final boolean empty = sk.getN() == 0;
+    final boolean lvlZeroSorted = sk.isLevelZeroSorted();
+    final boolean singleItem = sk.getN() == 1;
+    final boolean doubleType = (sk.sketchType == DOUBLES_SKETCH);
+    final int preInts = updatable
+        ? PREAMBLE_INTS_FULL
+        : (empty || singleItem) ? PREAMBLE_INTS_EMPTY_SINGLE : PREAMBLE_INTS_FULL;
+    //load the preamble
+    setMemoryPreInts(wmem, preInts);
+    final int server = updatable ? SERIAL_VERSION_UPDATABLE
+        : (singleItem ? SERIAL_VERSION_SINGLE : SERIAL_VERSION_EMPTY_FULL);
+    setMemorySerVer(wmem, server);
+    setMemoryFamilyID(wmem, Family.KLL.getID());
+    setMemoryEmptyFlag(wmem, empty);
+    setMemoryLevelZeroSortedFlag(wmem, lvlZeroSorted);
+    setMemorySingleItemFlag(wmem, singleItem);
+    setMemoryDoubleSketchFlag(wmem, doubleType);
+    setMemoryUpdatableFlag(wmem, updatable);
+    setMemoryK(wmem, sk.getK());
+    setMemoryM(wmem, sk.getM());
+  }
+
   /**
    * @param fmt format
    * @param args arguments
@@ -346,5 +985,5 @@ public class KllHelper {
   private static void println(final Object o) {
     System.out.println(o.toString());
   }
-}
 
+}
